requirements_dev.txt
jupyter==1.0.0
Pillow==9.5.0
notebook==6.5.4
numpy==1.23.5
opencv_python==4.7.0.72
ruff==0.0.263
streamlit==1.22.0
tensorflow==2.12.0
tensorflow-datasets==4.9.2